EntityUid Type (#498)
* Wraps Entity Uids in their own type.

* EntityId can't be serialized by entity params.

+++ SS14.Client/GameObjects/Components/Input/ClickableComponent.cs
@@ -27,7 +27,7 @@ namespace SS14.Client.GameObjects
         public void DispatchClick(IEntity user, int clickType)
         {
             OnClick?.Invoke(this, new ClickEventArgs(user, Owner, clickType));
-            Owner.SendComponentNetworkMessage(this, clickType, user.Uid);
+            Owner.SendComponentNetworkMessage(this, clickType, (int)user.Uid);
         }
     }
 }
